+++ src/plugin.js
@@ -5,6 +5,7 @@ var old = $.fn.bootgrid;
 
 $.fn.bootgrid = function (option)
 {
+    var args = Array.prototype.slice.call(arguments, 1);
     return this.each(function ()
     {
         var $this = $(this),
@@ -22,7 +23,7 @@ $.fn.bootgrid = function (option)
         }
         if (typeof option === "string")
         {
-            return instance[option]();
+            return instance[option].apply(instance, args);
         }
     });
 };